requirements.txt
rasterio==1.1.5
rio_tiler==1.3.1
Django==3.0.7
folium==0.11.0
numpy==1.18.2
matplotlib==3.3.0
uvicorn==0.11.8
fastapi==0.65.1
Pillow==8.2.0
aiofiles==0.7.0
gunicorn==20.0.4